+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/ValidatorVcJws.kt
@@ -75,6 +75,7 @@ class ValidatorVcJws(
         }
 
         val vp = VerifiablePresentationParsed(
+            jws = input,
             id = vpJws.vp.id,
             type = vpJws.vp.type,
             freshVerifiableCredentials = verificationResultWithFreshnessSummary.filter {