+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/data/IsoDocumentParsed.kt
@@ -1,5 +1,6 @@
 package at.asitplus.wallet.lib.data
 
+import at.asitplus.iso.Document
 import at.asitplus.iso.IssuerSignedItem
 import at.asitplus.iso.MobileSecurityObject
 import at.asitplus.wallet.lib.agent.validation.CredentialFreshnessSummary
@@ -9,6 +10,7 @@ import at.asitplus.wallet.lib.agent.validation.CredentialFreshnessSummary
  * and also in [at.asitplus.wallet.lib.agent.VerifierAgent.verifyPresentationIsoMdoc].
  */
 data class IsoDocumentParsed(
+    val document: Document,
     val mso: MobileSecurityObject,
     val validItems: List<IssuerSignedItem> = listOf(),
     val invalidItems: List<IssuerSignedItem> = listOf(),